Split input dispatch by mode
